+++ prototype/scripts/install-resource-broker-platform.js
@@ -1,0 +1,6 @@
+#!/usr/bin/env node
+'use strict';
+
+require(process.platform === 'win32'
+  ? './install-resource-broker-wsl'
+  : './install-resource-broker');